fix: 블랙잭인 경우 상금 1.5배 획득되지 않던 부분 수정

+++ src/main/java/blackjack/view/OutputView.java
@@ -76,8 +76,8 @@ public class OutputView {
         System.out.println("딜러 : " + dealerResult);
     }
 
-    public static void playersResult(Map<String, Integer> resultWithName) {
-        for (Map.Entry<String, Integer> entry : resultWithName.entrySet()) {
+    public static void playersResult(Map<String, Double> resultWithName) {
+        for (Map.Entry<String, Double> entry : resultWithName.entrySet()) {
             System.out.println(entry.getKey() + " : " + entry.getValue());
         }
     }
